working

+++ src/red4ext/IsAxis.cpp
@@ -5,5 +5,5 @@ using namespace RED4ext;
 
 bool IsAxis(uint16_t key) {
   // return ((key - 150) <= 5);
-  return ((key - 150) <= 5) || IsJoystickAxis(key);
+  return ((uint16_t)(key - 150U) <= 5U) || IsJoystickAxis(key);
 }
